requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
openai-whisper==20240930
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
langchain==0.3.4
langchain-core==0.3.34
langchain-ollama==0.2.0
gspread==6.2.0
python-dotenv==1.0.1
